Fix test (Power try get more then it need)
