+++ src/test/java/org/terasology/launcher/util/TestFileUtils.java
@@ -164,38 +164,6 @@ public class TestFileUtils {
     }
 
     @Test
-    public void testExtract() throws IOException {
-        final String fileInRoot = "fileInRoot";
-        final String fileInFolder = "folder/fileInFolder";
-        final String file1Contents = SAMPLE_TEXT + "1";
-        final String file2Contents = SAMPLE_TEXT + "2";
-        /* An archive with this structure is created:
-         * <zip root>
-         * +-- fileInRoot
-         * +-- folder
-         * |   +-- fileInFolder
-         */
-        Path zipFile = tempFolder.newFile(FILE_NAME + ".zip").toPath();
-        try (ZipOutputStream zipOutputStream = new ZipOutputStream(Files.newOutputStream(zipFile))) {
-            zipOutputStream.putNextEntry(new ZipEntry(fileInRoot));
-            zipOutputStream.write(file1Contents.getBytes());
-            zipOutputStream.closeEntry();
-            zipOutputStream.putNextEntry(new ZipEntry(fileInFolder));
-            zipOutputStream.write(file2Contents.getBytes());
-            zipOutputStream.closeEntry();
-        }
-
-        Path outputDir = tempFolder.newFolder().toPath();
-        FileUtils.extractZipTo(zipFile, outputDir);
-        Path extractedFileInRoot = outputDir.resolve(fileInRoot);
-        Path extractedFileInFolder = outputDir.resolve(fileInFolder);
-        assertTrue(Files.exists(extractedFileInRoot));
-        assertTrue(Files.exists(extractedFileInFolder));
-        assertEquals(file1Contents, Files.readAllLines(extractedFileInRoot).get(0));
-        assertEquals(file2Contents, Files.readAllLines(extractedFileInFolder).get(0));
-    }
-
-    @Test
     public void testDeleteFileSilently() throws IOException {
         Path tempFile = tempFolder.newFile(FILE_NAME).toPath();
         assertTrue(Files.exists(tempFile));
